Refactor to use the the exported classes in index.js
